third_party/blink/web_tests/external/wpt/WebCryptoAPI/import_export/ChaCha20-Poly1305_importKey.tentative.https.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, true, [encrypt])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, true, [decrypt, encrypt])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, true, [decrypt])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, true, [encrypt, decrypt, encrypt, decrypt])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, true, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, true, [encrypt])
  assert_unreached: Threw an unexpected error: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name Reached unreachable code
[FAIL] Good parameters: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, true, [decrypt, encrypt])
  assert_unreached: Threw an unexpected error: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name Reached unreachable code
[FAIL] Good parameters: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, true, [decrypt])
  assert_unreached: Threw an unexpected error: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name Reached unreachable code
[FAIL] Good parameters: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, true, [encrypt, decrypt, encrypt, decrypt])
  assert_unreached: Threw an unexpected error: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name Reached unreachable code
[FAIL] Empty Usages: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, true, [])
  assert_equals: Should throw correct error, not NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name expected "SyntaxError" but got "NotSupportedError"
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, false, [encrypt])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, false, [decrypt, encrypt])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, false, [decrypt])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, false, [encrypt, decrypt, encrypt, decrypt])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: ChaCha20-Poly1305}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, false, [encrypt])
  assert_unreached: Threw an unexpected error: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name Reached unreachable code
[FAIL] Good parameters: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, false, [decrypt, encrypt])
  assert_unreached: Threw an unexpected error: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name Reached unreachable code
[FAIL] Good parameters: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, false, [decrypt])
  assert_unreached: Threw an unexpected error: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name Reached unreachable code
[FAIL] Good parameters: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, false, [encrypt, decrypt, encrypt, decrypt])
  assert_unreached: Threw an unexpected error: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name Reached unreachable code
[FAIL] Empty Usages: 256 bits (jwk, {k: AQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHyA, kty: oct}, {name: ChaCha20-Poly1305}, false, [])
  assert_equals: Should throw correct error, not NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': Algorithm: Unrecognized name expected "SyntaxError" but got "NotSupportedError"
Harness: the test ran to completion.
